Edit func for create camera

+++ mac.js
@@ -7,8 +7,11 @@ const fs = require("fs");
 const ini = require('ini');
 
 let appName = "Electron";
-// let localData, fileName = "data-storage.log";
-async function Auth(username, password) {
+let configs = [];
+let pyData = [];
+let auth ;
+
+async function Auth(username) {
     try {
         // login logic => get username
         if(username.length === 0) {
@@ -26,26 +29,134 @@ async function Auth(username, password) {
    
 }
 
-async function postScanCamera(username, d) {
+// async function postScanCamera(username) {
+//     try {
+//         let url = `https://api-dev-revamp.viact.net/api/v2/cameras/detection/scan/create`
+//         let { status, data } = await axios.post(url, {username, data: auth});
+//         if(!status || status !== 201 ) {
+//             console.error('error:', error);
+//         }
+//         console.log("postScanCamera", data);
+//     } catch (error) {
+//         console.error('error:', error);
+//     }
+// }
+
+async function CreateCamera() {
+    let i = document.getElementById('config').value;
+    let port = document.getElementById('port').value;
+    let remote_port = document.getElementById('remote_port').value;
+    let usercamera = document.getElementById('usercamera').value;
+    let passwordcamera = document.getElementById('passwordcamera').value;
+    let postfix = document.getElementById('postfix').value;
+
+    console.log("CreateCamera", pyData[i]);
+    let item ={'remotePort': remote_port, 'localPort': port, 'localIp': pyData[i].ip};
+    configs.push(item)
+
+    // create forward
+    await createFrpcForward(item);
+
+    // create forward
+    let {URL} = await createCameraUrl({remotePort: remote_port, cameraVideoPath: postfix});
+    console.log("createCameraUrl URLURLURLURLURL:", URL);
+
+    // create camera (API củ)
+
+    // config frpc
+    await frpcClient(item);
+
+
+    // Note: dùng promise.all 
+}
+
+async function deleteCamera() {
+    // chosse camera
+
+    // remove item in array tổng (configs)
+
+    // delete forward
+    await deleteFrpcForward({remote_port: 'remote_port'});
+
+    // delete forward
+    await deleteCameraUrl({remotePort: 'remote_port'});
+
+    // delete camera (API củ)
+
+    // config frpc
+    await frpcClient(item);
+
+    // Note: dùng promise.all 
+}
+
+async function createFrpcForward(d) {
     try {
-        let url = `https://api-dev-revamp.viact.net/api/v2/cameras/detection/scan/create`
-        let { status, data } = await axios.post(url, {username, data: d});
+        let url = `https://api-dev-revamp.viact.net/api/v2/frpc-forwards`
+        let { status, data } = await axios.post(url, {
+            username: auth.username, 
+            remote_port: d.remotePort, 
+            local_port: d.localPort, 
+            local_ip: d.localIp
+        });
         if(!status || status !== 201 ) {
             console.error('error:', error);
         }
-        console.log("postScanCamera", data);
+        console.log("createFrpcForward");
     } catch (error) {
         console.error('error:', error);
     }
 }
 
-async function GetIP(username) {
-    console.log("scan..........");
-    document.getElementById("ip").innerHTML = "scan..........";
-
-    if(!username) {
-        username = document.getElementById('username').value;
+async function deleteFrpcForward(d) {
+    try {
+        let url = `https://api-dev-revamp.viact.net/api/v2/frpc-forwards/delete?username=${auth.username}&remote_port=${d.remote_port}`
+        let { status } = await axios.delete(url);
+        if(!status || status !== 200 ) {
+            console.error('error:', error);
+        }
+        console.log("deleteFrpcForward");
+    } catch (error) {
+        console.error('error:', error);
     }
+}
+
+async function createCameraUrl(d) {
+    try {
+        let url = `https://api.viact.net/cgi-config/camera-url`
+        let { status, data } = await axios.post(url, {
+            clientID: auth.username, 
+            remotePort:Number(d.remotePort), 
+            cameraVideoPath: d.cameraVideoPath
+        });
+        if(!status || status !== 200 ) {
+            console.error('error:', error);
+        }
+        console.log("createCameraUrl");
+        return data;
+    } catch (error) {
+        console.error('error:', error);
+    }
+}
+
+async function deleteCameraUrl(d) {
+    try {
+        let url = `https://api.viact.net/cgi-config/camera-url`
+        let { status, data } = await axios.delete(url, {
+            clientID: auth.username, 
+            remotePort:Number(d.remotePort), 
+        });
+        if(!status || status !== 200 ) {
+            console.error('error:', error);
+        }
+        console.log("deleteCameraUrl");
+        return data;
+    } catch (error) {
+        console.error('error:', error);
+    }
+}
+
+async function GetIP() {
+    document.getElementById("ip").innerHTML = "scan..........";
 
     var cmd = ""
     switch (process.platform) {
@@ -70,9 +181,9 @@ async function GetIP(username) {
     }
     console.log("cmdcmdcmdcmdcmdcmdcmd", cmd);
     const { stdout, stderr } = await exec(cmd);
-    console.log('stdout:', stdout);
     console.log('stderr:', stderr);
-    await postScanCamera(username, stdout);
+    eval('var arr='+stdout);
+    pyData = arr;
     document.getElementById("ip").innerHTML = stdout;
 }
 
@@ -117,12 +228,15 @@ function getAppDataPath() {
 
 async function InstallPackage() {
     let un = document.getElementById('username').value;
-    // let pw = document.getElementById('password').value;
-    let d = await Auth(un, "123");
-    await GetIP(d.username);
-    
+    let d = await Auth(un);
+    console.log(d);
+    auth = d;
+    await frpcClient(d);
+}
+
+async function frpcClient() {
     let pab = '';
-    let pai = createIni(d);
+    let {appDataFilePath,isReload} = createIni(auth);
 
     switch (process.platform) {
         case "darwin": {
@@ -145,35 +259,54 @@ async function InstallPackage() {
         }
     }
   
-    let cmd = `${pab} -c ${pai}`;
+    let cmd = `${pab} -c ${appDataFilePath}`;
+    if(isReload) {
+        cmd = `${pab} reload -c ${appDataFilePath}`;
+    }
     console.log('cmd:', cmd);
     const { stdout, stderr } = await exec(cmd);
     console.log('stdout:', stdout);
     console.log('stderr:', stderr);
 }
 
-function createIni(d) {
+function createIni() {
     const appDataDir = getAppDataPath();
-
+    let isReload = true
     if (!fs.existsSync(appDataDir)) {
         fs.mkdirSync(appDataDir);
     }
 
-    let pai = path.join(__dirname, 'frp', 'frpc.ini')
+    let pai = path.join(__dirname, '../frp', 'frpc.ini')
+    const appDataFilePath = path.join(appDataDir, "config_frpc.ini");
     var config = ini.parse(fs.readFileSync(pai, 'utf-8'))
 
-    config.common.server_addr = '13.213.164.203';
-    config.common.server_port = '7000';
-    config.common.user = d.username;
-    config.common.meta_token = d.token;
+    if(configs.length === 0 && !fs.existsSync(appDataFilePath)) {
+        isReload = false
+        config.common.server_addr = '13.213.164.203';
+        config.common.server_port = '7000';
+        config.common.user = auth.username;
+        config.common.meta_token = auth.token;
 
-    config.tcp_7000.remote_port = '2323';
-    config.tcp_7000.local_port = '7000';
-    config.tcp_7000.local_ip = '127.0.0.1';
+        fs.writeFileSync(appDataFilePath, ini.stringify(config));
+    }
 
-    const appDataFilePath = path.join(appDataDir, "config_frpc.ini");
+    if(configs.length > 0 && fs.existsSync(appDataFilePath)) {
+        config.common.server_addr = '13.213.164.203';
+        config.common.server_port = '7000';
+        config.common.user = auth.username;
+        config.common.meta_token = auth.token;
 
-    fs.writeFileSync(appDataFilePath, ini.stringify(config));
+        for(let i=0; i < configs.length; i++) {
+            config[`item_${i}`]= {
+                type: 'tcp',
+                remote_port: configs[i].remotePort,
+                local_port: configs[i].localPort,
+                local_ip: configs[i].localIp,
+            };
+        }
 
-    return appDataFilePath
+        fs.writeFileSync(appDataFilePath, ini.stringify(config));
+    }
+
+    return {appDataFilePath,isReload}
 }
